fix: gpt helpers
moved base ask and context call into helpers
